# go/model_articles_response.go
/*
 * simple blog
 *
 * A Simple Blog
 *
 * API version: 1.0.0
 * Contact: apiteam@swagger.io
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */

package swagger

type ArticlesResponse struct {

	Articles []ArticleResponse `json:"Articles,omitempty"`
}